Gulpfile for monorepo
