+++ packages/cli/index.js
@@ -1,0 +1,5 @@
+const { CLI } = require('./lib')
+
+const cli = CLI.new()
+
+cli.run()
