+++ packages/cli/src/serve.ts
@@ -16,7 +16,7 @@ import { Subject } from 'rxjs'
 
 import { Command, CommandOptions } from './command'
 import { Compiler } from '@posttext/compiler'
-import { getInterpreters } from '@posttext/printer/web'
+import { getInterpreters } from '@posttext/printer'
 import { Logger } from './helpers/logger'
 
 export class ServeCommand implements Command {
