+++ packages/cli/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": "src",
     "module": "commonjs",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
@@ -9,6 +10,5 @@
     "alwaysStrict": true,
     "strictNullChecks": true,
     "noUnusedLocals": true
-  },
-  "exclude": ["dist", "lib"]
+  }
 }
