+++ packages/printer/src/ast.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import * as ast from '@posttext/printer/ast'
+import * as ast from '@posttext/parser'
 
 export interface DocumentNode extends ast.DocumentNode {
   __metadata: Record<string, any>
