+++ packages/printer/src/index.ts
@@ -4,3 +4,5 @@
 
 export * from './command'
 export * from './printer'
+
+export * from './web'
