+++ packages/registry/src/resolver.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { Command } from '@posttext/printer/command'
+import { Command } from '@posttext/printer'
 
 export interface Resolver {
   preload?(): AsyncGenerator<Command, void, any>
